Invert the spectacular → inimitable and spectacular → digression edges
`Uuid is Showable` and `StackTrace is Showable` were the only reason
`spectacular` depended on `inimitable` and `digression`. Each moves into its
subject type's own companion, which is the same implicit scope, so no call site
needs an import; the build edge flips in the same commit and `spectacular.core`
is left with stenography and wisteria.

`StackTrace`'s renderer used `" "*n`, which resolved to a proscenium extension
in spectacular but to symbolism's `Multiplicable` in digression, where it has no
instance for String. It now uses `String.repeat`, which is identical for the
non-negative counts these are by construction (each is a maximum width minus a
member's length).

The roadmap's rejected item stands: `typeRepr` and `meta` stay in `Showable`'s
companion, since compiler-owned types have no companion to host them and moving
them would change `summonFrom` fallthrough.
